# Optimize IR expression data structures