--- a/fmm/src/ir/bitwise_not_operation.rs
+++ b/fmm/src/ir/bitwise_not_operation.rs
@@ -3,24 +3,30 @@ use crate::types;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct BitwiseNotOperation {
+pub struct BitwiseNotOperation(Arc<BitwiseNotOperationInner>);
+
+#[derive(Clone, Debug, PartialEq)]
+struct BitwiseNotOperationInner {
     type_: types::Primitive,
-    value: Arc<Expression>,
+    value: Expression,
 }
 
 impl BitwiseNotOperation {
     pub fn new(type_: types::Primitive, value: impl Into<Expression>) -> Self {
-        Self {
-            type_,
-            value: Arc::new(value.into()),
-        }
+        Self(
+            BitwiseNotOperationInner {
+                type_,
+                value: value.into(),
+            }
+            .into(),
+        )
     }
 
     pub fn type_(&self) -> types::Primitive {
-        self.type_
+        self.0.type_
     }
 
     pub fn value(&self) -> &Expression {
-        &self.value
+        &self.0.value
     }
 }