--- a/fmm/src/ir/union.rs
+++ b/fmm/src/ir/union.rs
@@ -3,7 +3,10 @@ use crate::types;
 use std::sync::Arc;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Union {
+pub struct Union(Arc<UnionInner>);
+
+#[derive(Clone, Debug, PartialEq)]
+struct UnionInner {
     type_: types::Union,
     member_index: usize,
     member: Arc<Expression>,
@@ -11,22 +14,25 @@ pub struct Union {
 
 impl Union {
     pub fn new(type_: types::Union, member_index: usize, member: impl Into<Expression>) -> Self {
-        Self {
-            type_,
-            member_index,
-            member: member.into().into(),
-        }
+        Self(
+            UnionInner {
+                type_,
+                member_index,
+                member: member.into().into(),
+            }
+            .into(),
+        )
     }
 
     pub fn type_(&self) -> &types::Union {
-        &self.type_
+        &self.0.type_
     }
 
     pub fn member_index(&self) -> usize {
-        self.member_index
+        self.0.member_index
     }
 
     pub fn member(&self) -> &Expression {
-        &self.member
+        &self.0.member
     }
 }