--- a/fmm/src/ir/union.rs
+++ b/fmm/src/ir/union.rs
@@ -9,7 +9,7 @@ pub struct Union(Arc<UnionInner>);
 struct UnionInner {
     type_: types::Union,
     member_index: usize,
-    member: Arc<Expression>,
+    member: Expression,
 }
 
 impl Union {