--- a/fmm/src/ir/variable.rs
+++ b/fmm/src/ir/variable.rs
@@ -1,11 +1,15 @@
+use std::sync::Arc;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Variable {
-    name: String,
+    name: Arc<str>,
 }
 
 impl Variable {
     pub fn new(name: impl Into<String>) -> Self {
-        Self { name: name.into() }
+        Self {
+            name: name.into().into(),
+        }
     }
 
     pub fn name(&self) -> &str {